Split out blueprint
