+++ broadbean/__init__.py
@@ -4,3 +4,4 @@ from . import ripasso
 from .broadbean import *
 from .element import Element
 from .sequence import Sequence
+from .blueprint import BluePrint
